% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_direction_string.R
\name{apply_direction_string}
\alias{apply_direction_string}
\title{Apply the direction to the decimal coordinate within a string}
\usage{
apply_direction_string(string_decimal, string_direction)
}
\arguments{
\item{string_decimal}{numeric}

\item{string_direction}{minus or plus}
}
\value{
string
}
\description{
Apply the direction to the decimal coordinate within a string
}
