% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_direction_vector.R
\name{apply_direction_vector}
\alias{apply_direction_vector}
\title{Apply the direction to the decimal coordinate within coordinate vector}
\usage{
apply_direction_vector(string_decimal, string_direction)
}
\arguments{
\item{string_decimal}{numeric}

\item{string_direction}{plus or minus}
}
\value{
vector with applied direction vector
}
\description{
Apply the direction to the decimal coordinate within coordinate vector
}
