% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coord_data}
\alias{coord_data}
\title{Coordinate Example Data}
\format{
\subsection{\code{coord_data}}{

A data frame with 1,668 rows and 8 columns:
\describe{
\item{Plot_ID}{Unique Plot ID}
\item{DataSource_ID}{Unique Datasource ID}
\item{Location}{Location where the measurement was taken}
\item{Latitude}{Latitude}
\item{Longitude}{Longitude}
\item{StartYear}{Start of the experiment}
\item{EndYear}{End of the experiment}
\item{Country}{Country of the experiment}
}
}
}
\source{
generated
}
\usage{
coord_data
}
\description{
A dummy dataset that contains unorganized coordinates to experiment with the package
}
\keyword{datasets}
