% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_digits_from_string.R
\name{extract_digits_from_string}
\alias{extract_digits_from_string}
\title{Extract digits from a string after replace_nondigits_string/vecotr was applied}
\usage{
extract_digits_from_string(string)
}
\arguments{
\item{string}{is an unorganized string}
}
\value{
list with first second and third digits saved as elements
}
\description{
Extract digits from a string after replace_nondigits_string/vecotr was applied
}
