% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_digits_from_vector.R
\name{extract_digits_from_vector}
\alias{extract_digits_from_vector}
\title{Extract digits from a vector after replace_nondigits_string/vecotr was applied}
\usage{
extract_digits_from_vector(strings)
}
\arguments{
\item{strings}{}
}
\value{
vector of lists aka a column
}
\description{
Extract digits from a vector after replace_nondigits_string/vecotr was applied
}
