% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_direction_string.R
\name{extract_direction_string}
\alias{extract_direction_string}
\title{Extract direction (East/South/West/North) as a minus or plus from a string}
\usage{
extract_direction_string(coordinate_string)
}
\arguments{
\item{coordinate_string}{is either latitude or longitude column to which the function is applied}
}
\value{
returns a coordinate string for further processing with extract_direction_vector
}
\description{
Extract direction (East/South/West/North) as a minus or plus from a string
}
\examples{
extract_direction_string("34.56E")
extract_direction_string("34.56W")
}
