% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_direction_vector.R
\name{extract_direction_vector}
\alias{extract_direction_vector}
\title{Extract direction (East/South/West/North) as a minus or plus from a vector}
\usage{
extract_direction_vector(strings)
}
\arguments{
\item{strings}{is a column in a dataset}
}
\value{
returns a vector (column) which contains directions of the coordinate string (plus or minus)
}
\description{
Extract direction (East/South/West/North) as a minus or plus from a vector
}
