% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_coordinate_format.R
\name{identify_coordinate_format}
\alias{identify_coordinate_format}
\title{Identify coordinate format}
\usage{
identify_coordinate_format(df, longitude, latitude)
}
\arguments{
\item{df}{a dataset}

\item{longitude}{unorganised string}

\item{latitude}{unorganised string}
}
\value{
a dataset with 2 new columns for long and lat formats
}
\description{
This function takes a dataset as an input and then adds another column with longitude and latitude corrected.
}
\examples{
identify_coordinate_format(
df = coord_data,
longitude = "Longitude",
latitude = "Latitude"
)
}
