% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_nondigits_string.R
\name{replace_nondigits_string}
\alias{replace_nondigits_string}
\title{Replace non-digits within a string}
\usage{
replace_nondigits_string(string)
}
\arguments{
\item{string}{is an unorganized string}
}
\value{
string with replaced nondigits
}
\description{
Replace non-digits within a string
}
