% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_nondigits_vector.R
\name{replace_nondigits_vector}
\alias{replace_nondigits_vector}
\title{Replace non-digits in a vector}
\usage{
replace_nondigits_vector(strings)
}
\arguments{
\item{strings}{is a column}
}
\value{
vector (aka column in a dataset) with replaced nondigits
}
\description{
Replace non-digits in a vector
}
