% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_coordinates_to_dd.R
\name{process_coordinates_to_dd}
\alias{process_coordinates_to_dd}
\title{Process the dataset and covert all types of coordinates to decimal degrees.}
\usage{
process_coordinates_to_dd(df, longitude, latitude)
}
\arguments{
\item{df}{is a dataset}

\item{longitude}{its longitude column}

\item{latitude}{its latitude column}
}
\value{
a dataset with new columns with coordinates in decimal format
}
\description{
Process the dataset and covert all types of coordinates to decimal degrees.
}
\examples{
process_coordinates_to_dd(
df = coord_data,
longitude = "Longitude",
latitude = "Latitude")

}
