% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorten_ddm_intervals_string.R
\name{shorten_ddm_intervals_string}
\alias{shorten_ddm_intervals_string}
\title{Shorten DDM intervals within a string}
\usage{
shorten_ddm_intervals_string(coordinate_strings, option = "first")
}
\arguments{
\item{coordinate_strings}{is a coordinate column}

\item{option}{"first"(the part before the "-" in the interval) or "second" (the part after the "-" in the interval)}
}
\value{
coordinate string with where interval is cut to the first or second part
}
\description{
Shorten DDM intervals within a string
}
